code/back/api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Coffee Inventory Planning API"
requires-python = ">=3.11,<3.13"
dependencies = [
    "cplex>=22.1.2.0",
    "docplex>=2.30.251",
    "fastapi[standard]>=0.104.0",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.7",
    "numpy>=2.3.5",
    "pandas>=2.3.3",
    "plotly>=6.5.0",
    "uvicorn[standard]>=0.24.0",
    # Database & ORM
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "asyncpg>=0.29.0",
    # Authentication
    "python-jose[cryptography]>=3.3.0",
    "passlib>=1.7.4",
    "bcrypt>=4.0.0,<4.1.0",  # Pin bcrypt for passlib compatibility
    # Configuration
    "pydantic-settings>=2.0.0",
]

[tool.setuptools]
py-modules = []

[tool.setuptools.packages.find]
include = ["model*", "models*", "routers*", "schemas*", "services*", "middleware*", "repositories*", "db_handle*"]

[dependency-groups]
dev = [
    "mypy>=1.18.2",
    "ruff>=0.14.6",
]

[tool.mypy]
exclude = [".venv"]
disallow_untyped_defs = true
disallow_any_unimported = false
no_implicit_optional = true
check_untyped_defs = true
warn_return_any = true
warn_unused_ignores = true
show_error_codes = true
plugins = [
  "pydantic.mypy"
]

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]
target-version = "py312"
indent-width = 4
line-length = 120

[tool.ruff.lint]
# By default, Ruff enables Flake8's E and F rules
# Pyflakes - F, pycodestyle - E, W
# flake8-builtins - A
# Pylint - PLC, PLE, PLW
# isort - I
select = ['E', 'F', 'W', 'A', 'PLC', 'PLE', 'PLW', 'I', 'FAST']
ignore = [
    # LineTooLong
    "E501",
    # DoNotAssignLambda
    "E731",
]
# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"
